+++ tests/test_NIRSpec.py
@@ -15,7 +15,7 @@ from eureka.S4_generate_lightcurves import s4_genLC as s4
 from eureka.S5_lightcurve_fitting import s5_fit as s5
 
 
-def test_NIRSpec(capsys, keep_s2_output, keep_s3_output):
+def test_NIRSpec(capsys, keep_s2_output, keep_s3_output, keep_s4_output):
     with capsys.disabled():
         # is able to display any message without failing a test
         # useful to leave messages for future users who run the tests
@@ -99,5 +99,6 @@ def test_NIRSpec(capsys, keep_s2_output, keep_s3_output):
                   f"Stage2{os.sep}S2_*")
     if not keep_s3_output:
         os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRSpec{os.sep}Stage3")
-    os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRSpec{os.sep}Stage4")
+    if not keep_s4_output:
+        os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRSpec{os.sep}Stage4")
     os.system(f"rm -r data{os.sep}JWST-Sim{os.sep}NIRSpec{os.sep}Stage5")
