+++ tests/test_Photometry_NIRCam.py
@@ -14,7 +14,7 @@ from eureka.S4_generate_lightcurves import s4_genLC as s4
 from eureka.S5_lightcurve_fitting import s5_fit as s5
 
 
-def test_NIRCamPhotometry(capsys, keep_s3_output):
+def test_NIRCamPhotometry(capsys, keep_s3_output, keep_s4_output):
     with capsys.disabled():
         # is able to display any message without failing a test
         # useful to leave messages for future users who run the tests
@@ -76,7 +76,8 @@ def test_NIRCamPhotometry(capsys, keep_s3_output):
     # remove temporary files
     if not keep_s3_output:
         os.system(f"rm -r data{os.sep}Photometry{os.sep}NIRCam{os.sep}Stage3")
-    os.system(f"rm -r data{os.sep}Photometry{os.sep}NIRCam{os.sep}Stage4")
+    if not keep_s4_output:
+        os.system(f"rm -r data{os.sep}Photometry{os.sep}NIRCam{os.sep}Stage4")
     os.system(f"rm -r data{os.sep}Photometry{os.sep}NIRCam{os.sep}Stage5")
